feat(NavUser): refactor user information display by extracting UserInfo component for better code organization and reusability fix(NavUser): update dropdown menu items and labels for clarity and consistency in user actions
